ModelLearning/item-detection/pyproject.toml
[tool.poetry]
name = "item-detection"
version = "0.1.0"
description = ""
authors = ["JakubJakacki18 <jakub.jakacki18@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
poetry-core = "^2.1.3"
python = "^3.9,<3.10"
tensorflow = "2.8.4"
tflite-model-maker = "0.4.3"
numpy = "1.21.6"
pandas = "1.3.5"
array-record = "0.5.0"
pycocotools = "^2.0.10"
tflite-support = "^0.4.4"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

